Sheets API Call working
I had some issues with massaging strings into no-spaces and all-caps for IDs, but everything seems to be working now. I'm really happy with how the layout is working. Right now, when the data is fetched and creates HTML, the website scrolls down, which I would like it not to. Not sure where to start with this issue.